Restructure WikidataStuff (#52)
Break out `Reference`, `Qualifier` and `Statement` from `WikidataStuff`.

Simplify dependencies by disconnecting `helpers` from other classes (using
`@deprecated` for backwards compatibility).
* Move `listify` and `list_to_lower` from `WikidataStuff` to `helpers`.
* Move `add_start_end_qualifiers` from `helpers` to `Statement`.

Also moves the declaration of `pywikibot.Claim.__repr__` from
`WikidataStuff.py` to `__init__.py`.

Closes: #22 
Bug: [T205538](https://phabricator.wikimedia.org/T205538)
