+++ wikidataStuff/WikidataStuff.py
@@ -12,221 +12,15 @@ import os.path  # Needed for WikidataStringSearch
 import pywikibot
 from pywikibot.tools import deprecated, deprecated_args
 
+import wikidataStuff.helpers as helpers
+from wikidataStuff.reference import Reference
+
 
 class WikidataStuff(object):
     """Useful methods for interacting with Wikidata using pywikibot."""
 
     repo = None
     edit_summary = None
-
-    # extend pywikibot.Claim with a __repr__ method
-    def new_repr(self):
-        """Override the normal representation of pywikibot.Claim."""
-        return 'WD.Claim(%s: %s)' % (self.getID(), self.getTarget())
-    pywikibot.Claim.__repr__ = new_repr
-
-    class Reference(object):
-        """
-        A class for encoding the contents of a reference.
-
-        Makes a distinction between the elements which should be included in a
-        comparison with other references and those which shouldn't.
-
-        e.g. for "reference URL: some URL", "retrieved: some_date" you would
-        want to compare sources on the URL but not the date.
-
-        A comparison will fail if ANY of the source_test sources are present.
-        """
-
-        def __init__(self, source_test=None, source_notest=None):
-            """
-            Make a Reference object from the provided sources.
-
-            @param source_test: claims which should be included in
-                comparison tests
-            @type source_test: pywikibot.Claim|list of pywikibot.Claim
-            @param source_notest: claims which should be excluded from
-                comparison tests
-            @type source_notest: pywikibot.Claim|list of pywikibot.Claim
-            """
-            # avoid mutable default arguments
-            source_test = source_test or []
-            source_notest = source_notest or []
-
-            # standardise the two types of allowed input
-            self.source_test = listify(source_test)
-            self.source_notest = listify(source_notest)
-
-            # validate input
-            self.validate_sources()
-
-        def validate_sources(self):
-            """Validate the sources of a Reference."""
-            sources = self.get_all_sources()
-            if not sources:
-                raise pywikibot.Error(
-                    'You tried to create a reference without any sources')
-            if not all(isinstance(s, pywikibot.Claim) for s in sources):
-                raise pywikibot.Error(
-                    'You tried to create a reference with a non-Claim source')
-
-        def get_all_sources(self):
-            """Return all the sources of a Reference."""
-            return self.source_test + self.source_notest
-
-        def __repr__(self):
-            """Return a more complete string representation."""
-            return 'WD.Reference(test: %s, no_test: %s)' % (
-                self.source_test, self.source_notest)
-
-    class Qualifier(object):
-        """
-        A class for encoding the contents of a qualifier.
-
-        Essentially pywikibot.Claim without having to provide an instantiated
-        repo.
-
-        @todo: redo as SimpleClaim (if so reuse in Reference) or
-               retire in favor of pywikibot.Claim
-        """
-
-        def __init__(self, P, itis):
-            """
-            Make a correctly formatted qualifier object for claims.
-
-            @param P: the property (with or without "P")
-            @type P: basestring
-            @param itis: a valid claim target e.g. pywikibot.ItemPage
-            @type itis: object
-            """
-            self.prop = 'P%s' % str(P).lstrip('P')
-            self.itis = itis
-
-        def __repr__(self):
-            """Return a more complete string representation."""
-            return 'WD.Qualifier(%s, %s)' % (self.prop, self.itis)
-
-        def __eq__(self, other):
-            """Implement equality comparison."""
-            if isinstance(other, self.__class__):
-                return self.__dict__ == other.__dict__
-            return NotImplemented
-
-        def __ne__(self, other):
-            """Implement non-equality comparison."""
-            return not self.__eq__(other)
-
-        def __hash__(self):
-            """Implement hash to allow for e.g. sorting and sets."""
-            return hash((self.prop, self.itis))
-
-    class Statement(object):
-        """A representation of a statement (value, qualifiers, references)."""
-
-        def __init__(self, itis, special=False):
-            """
-            Make a correctly formatted statement object for claims.
-
-            @todo: itis test
-
-            @param itis: a valid claim target e.g. pywikibot.ItemPage
-            @type itis: object
-            @param special: if itis is actually a snakvalue
-            @type special: bool
-            """
-            if special and itis not in ['somevalue', 'novalue']:
-                raise pywikibot.Error(
-                    'You tried to create a special statement with a '
-                    'non-allowed snakvalue: %s' % itis)
-            self.itis = itis
-            self._quals = set()
-            self.ref = None
-            self.special = special
-            self.force = False
-
-        def addQualifier(self, qual, force=False):
-            """
-            Add qualifier to the statement if not None or already present.
-
-            Returns self to allow chaining.
-
-            @param qual: the qualifier to add
-            @type qual: Qualifier|None
-            @param force: whether qualifier should be added even to already
-                sourced items
-            @type force: bool
-            @rtype Statement
-            """
-            # test input
-            if qual is None:
-                # simply skip any action
-                return self
-            elif not isinstance(qual, WikidataStuff.Qualifier):
-                raise pywikibot.Error(
-                    'addQualifier was called with something other '
-                    'than a Qualifier|None object: %s' % qual)
-
-            # register qualifier
-            self._quals.add(qual)
-            if force:
-                self.force = True
-            return self
-
-        def add_reference(self, ref):
-            """
-            Add a Reference to the statement.
-
-            Returns self to allow chaining.
-
-            @param ref: the reference to add
-            @type ref: Reference
-            @raises: pywikibot Error if a reference is already present or if
-                the provided value is not a Reference.
-            """
-            # test input
-            if self.ref is not None:
-                raise pywikibot.Error(
-                    'add_reference was called when the statement already had '
-                    'a reference assigned to it.')
-            elif not isinstance(ref, WikidataStuff.Reference):
-                raise pywikibot.Error(
-                    'add_reference was called with something other '
-                    'than a Reference object: %s' % ref)
-            else:
-                self.ref = ref
-
-            return self
-
-        def isNone(self):
-            """Test if Statement was created with itis=None."""
-            return self.itis is None
-
-        @property
-        def quals(self):
-            """Return the list of qualifiers."""
-            return list(self._quals)
-
-        def __repr__(self):
-            """Return a more complete string representation."""
-            return ('WD.Statement('
-                    'itis:{}, quals:{}, ref:{}, special:{}, force:{})'.format(
-                        self.itis, self.quals, self.ref, self.special,
-                        self.force))
-
-        def __eq__(self, other):
-            """Two Statements are equal if same up to qualifier order."""
-            if isinstance(other, self.__class__):
-                return self.__dict__ == other.__dict__
-            return NotImplemented
-
-        def __hash__(self):
-            """Implement hash to allow for e.g. sorting and sets."""
-            return hash((self.itis, frozenset(self._quals), self.ref,
-                         self.special, self.force))
-
-        def __ne__(self, other):
-            """Implement non-equality comparison."""
-            return not self.__eq__(other)
 
     def __init__(self, repo, edit_summary=None, no_wdss=False):
         """
@@ -312,7 +106,7 @@ class WikidataStuff(object):
             pywikibot.output(edit_summary)
 
     # @todo: deprecate in favour of add_label_or_alias
-    @deprecated_args(caseSensitive='case_sensitive')
+    @deprecated_args(caseSensitive='case_sensitive', since='<0.4')
     def addLabelOrAlias(self, lang, name, item, summary=None,
                         case_sensitive=False):
         """
@@ -359,7 +153,7 @@ class WikidataStuff(object):
         aliases = item.aliases or dict()
 
         for lang, names in data.items():
-            for name in listify(names):
+            for name in helpers.listify(names):
                 if lang not in labels:
                     labels[lang] = name
                     new_label_langs.append(lang)
@@ -371,7 +165,7 @@ class WikidataStuff(object):
                         new_alias_langs.append(lang)
                     elif (case_sensitive and name not in aliases[lang]) or \
                             (not case_sensitive and
-                             name.lower() not in list_to_lower(aliases[lang])):
+                             name.lower() not in helpers.list_to_lower(aliases[lang])):
                         aliases[lang].append(name)
                         new_alias_langs.append(lang)
 
@@ -525,7 +319,7 @@ class WikidataStuff(object):
                     'Something went very wrong trying to add a qualifier: %s' %
                     e)
 
-    @deprecated('hasClaim')
+    @deprecated('hasClaim', since='<0.4')
     def has_claim(self, prop, itis, item):
         """
         Check if the claim already exists, if so returns any matching claim.
@@ -610,7 +404,7 @@ class WikidataStuff(object):
             ref = statement.ref
 
         # test reference (must be a Reference or explicitly missing)
-        if not isinstance(ref, WikidataStuff.Reference) and ref is not None:
+        if not isinstance(ref, Reference) and ref is not None:
             raise pywikibot.Error('The provided reference was not a Reference '
                                   'object. Crashing')
 
@@ -825,31 +619,3 @@ class WikidataStuff(object):
 
         # return the new item
         return result
-
-
-# generic methods which are used already here and so could not be in helpers
-# since that would cause a cyclical import
-def list_to_lower(string_list):
-    """
-    Convert every string in a list to lower case.
-
-    @param string_list: list of strings to convert
-    @type string_list: list (of basestring)
-    @rtype: list (of basestring)
-    """
-    return [s.lower() for s in string_list]
-
-
-def listify(value):
-    """
-    Turn the given value, which might or might not be a list, into a list.
-
-    @param value: The value to listify
-    @rtype: list|None
-    """
-    if value is None:
-        return None
-    elif isinstance(value, list):
-        return value
-    else:
-        return [value, ]
